authfe: Log orgID in launcherRequests
